requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
python-dotenv==1.0.1
requests==2.32.3
jinja2==3.1.4
python-multipart==0.0.12